requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
coloredlogs==15.0.1
dnspython==2.7.0
eventlet==0.37.0
Flask==3.0.3
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
greenlet==3.1.1
h11==0.14.0
humanfriendly==10.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mpmath==1.3.0
numpy==2.1.2
onnx==1.17.0
onnxruntime==1.19.2
packaging==24.1
pillow==11.0.0
protobuf==5.28.2
python-engineio==4.10.1
python-socketio==5.11.4
simple-websocket==1.1.0
sympy==1.13.3
Werkzeug==3.0.4
wsproto==1.2.0
